Updated merged networkGraph to include scss and className
